Add once event listeners
I noticed that the types for event listeners do not include the `once` types alongside them, which keep the event listener attached for a single firing of the event before removing the handler automatically.

I've added these `once` variants of the `on` handler types. I did consider making a new type for each `on` handler and referencing it in the `once` variant, but I believe that would make things a lot more complex, and considering the amount of events and listeners in this library, would be counter-productive and wouldn't really increase maintainability by much. So instead I have simply cloned each `on` listener method and renamed them to `once`.